Move attacking from AttackableUnit to ObjAIBase
