+++ GameServerLib/Logic/GameObjects/AttackableUnits/AI/Champion.cs
@@ -352,11 +352,6 @@ namespace LeagueSandbox.GameServer.Logic.GameObjects
             return hash;
         }
 
-        public override bool isInDistress()
-        {
-            return DistressCause != null;
-        }
-
         public short getSkillPoints()
         {
             return _skillPoints;
